+++ R/SummariseBudget.R
@@ -20,7 +20,7 @@
 #'   Data in the MODFLOW cell-by-cell budget file must be saved using the
 #'   \emph{\bold{"COMPACT BUDGET"}} output option.
 #'
-#' @return Returns a 'data.frame' with the following variables:
+#' @return Returns a 'data.table' with the following variables:
 #'   \describe{
 #'     \item{desc}{description of data type, such as "wells".}
 #'     \item{kper}{stress period}
@@ -44,6 +44,8 @@
 #'
 #' @keywords utilities
 #'
+#' @importFrom data.table data.table
+#'
 #' @export
 #'
 #' @examples
@@ -59,15 +61,15 @@ SummariseBudget <- function(budget, desc=NULL, id=NULL) {
     checkmate::assertFileExists(budget)
     budget <- ReadModflowBinary(budget, "flow")
   }
-  checkmate::assertCharacter(desc, any.missing=FALSE, min.len=1, unique=TRUE, null.ok=TRUE)
+  checkmate::assertCharacter(desc, any.missing=FALSE, min.len=1,
+                             unique=TRUE, null.ok=TRUE)
   checkmate::assertString(id, null.ok=TRUE)
 
   budget.desc <- vapply(budget, function(i) i$desc, "")
-  if (is.null(desc)) {
-    desc <- unique(budget.desc)
-  } else {
+  if (!is.null(desc)) {
     is <- desc %in% budget.desc
-    if (all(!is)) stop("data type(s) not found in budget")
+    if (all(!is))
+      stop("data type(s) not found in budget")
     if (any(!is))
       warning(sprintf("data type(s) not found in budget: %s",
                       paste(paste0("\"", desc[!is], "\""), collapse=", ")))
@@ -75,56 +77,44 @@ SummariseBudget <- function(budget, desc=NULL, id=NULL) {
   }
 
   is <- vapply(budget, function(x) !is.null(colnames(x$d)), FALSE)
-  if (all(!is)) stop("data type(s) not saved using correct form")
+  if (all(!is))
+    stop("data type(s) not saved using correct form")
   if (any(!is)) {
     x <- unique(vapply(budget[!is], function(i) i$desc, ""))
     warning(sprintf("removed data type(s): %s not saved using correct form",
-            paste(paste0("\"", x, "\""), collapse=", ")))
+                    paste(paste0("\"", x, "\""), collapse=", ")))
   }
   budget <- budget[is]
-  desc <- vapply(budget, function(i) i$desc, "")
 
-  b <- budget
-  for (i in seq_along(b)) b[[i]]$d[b[[i]]$d[, "flow"] < 0, "flow"] <- 0
-  d <- dplyr::mutate(.Summarise(b, desc, id), flow.dir="in")
-
-  b <- budget
-  for (i in seq_along(b)) b[[i]]$d[b[[i]]$d[, "flow"] > 0, "flow"] <- 0
-  d <- dplyr::bind_rows(d, dplyr::mutate(.Summarise(b, desc, id), flow.dir="out"))
-
-  d$flow.dir <- as.factor(d$flow.dir)
-  d
-}
-
-
-.Summarise <- function(b, desc, id) {
-
-  checkmate::assertList(b)
-  checkmate::assertCharacter(desc)
-  checkmate::assertString(id, null.ok=TRUE)
-
-  d <- dplyr::bind_rows(lapply(desc, function(i) {
-    dplyr::bind_rows(lapply(b[desc == i], function(j) {
-      d <- data.frame(desc=j$desc, kper=j$kper, kstp=j$kstp, id=NA,
-                      flow=j$d[, "flow"], delt=j$delt,
-                      pertim=j$pertim, totim=j$totim, stringsAsFactors=FALSE)
-      if (!is.null(id) && id %in% colnames(j$d)) d$id <- j$d[, id]
-      d
-    }))
+  dt <- data.table::rbindlist(lapply(budget, function(x) {
+    d <- data.table::data.table(desc=x$desc, kper=x$kper, kstp=x$kstp, id=NA,
+                                flow=x$d[, "flow"], delt=x$delt,
+                                pertim=x$pertim, totim=x$totim)
+    if (!is.null(id) && id %in% colnames(x$d)) d$id <- x$d[, id]
+    d
   }))
-  d$desc <- as.factor(d$desc)
-  if ("id" %in% colnames(d))
-    grps <- dplyr::group_by_(d, "desc", "kper", "kstp", "id")
-  else
-    grps <- dplyr::group_by_(d, "desc", "kper", "kstp")
-  d <- dplyr::summarise_(grps,
-                         delt        = "delt[1]",
-                         pertim      = "pertim[1]",
-                         totim       = "totim[1]",
-                         count       = "length(flow)",
-                         flow.sum    = "sum(flow)",
-                         flow.mean   = "mean(flow)",
-                         flow.median = "stats::median(flow)",
-                         flow.sd     = "sd(flow)")
-  d
+  dt$desc <- as.factor(dt$desc)
+  dt$flow.dir <- as.character(NA)
+
+  # due to NSE notes in R CMD check
+  delt <- flow <- flow.dir <- kper <- kstp <- pertim <- totim <- NULL
+
+  dt_summary <- data.table::rbindlist(lapply(c("in", "out"), function(i) {
+    if (i == "in")
+      dt$flow[dt$flow < 0] <- 0
+    else
+      dt$flow[dt$flow > 0] <- 0
+    dt[, list(delt        = utils::head(delt, 1),
+              pertim      = utils::head(pertim, 1),
+              totim       = utils::head(totim, 1),
+              count       = length(flow),
+              flow.sum    = sum(flow),
+              flow.mean   = mean(flow),
+              flow.median = stats::median(flow),
+              flow.sd     = stats::sd(flow),
+              flow.dir    = i),
+       by=list(desc, kper, kstp, id)]
+  }))
+  dt_summary$flow.dir <- as.factor(dt_summary$flow.dir)
+  dt_summary
 }